Update to allow processing of small samples
If the number of cells is < 200 for a sample, we need to reduce the default values. Seurat expects at least 200 cells per sample for the generate_scrna_integration_seurat.

If the number of cells is < 200 for any sample, we set the k.filter and k.weight values to the number cells the smallest sample has. 
This way, we can still integrate very small data sets.