- Standardized many of the plots in the report so they look more alike.
- Made a number of changes to the sizes of the plots, mainly so that these
have dynamic height, scaling with the number of entries/rows of the matrix or
data.frame from which they are created. In the cases where several plots
are put next to each other, the size of the plot with the most rows
takes precedence.

- For every enrichment term plot, added a line beyong which terms are
significant.

- Made a number of code style changes so that the code is easier to read,
might also be relevant if this repository is every turned into a fully
fledged package.

- Also connected to the last point, refactored some of the viz code so that
there is less code duplication (there is still much to work on though).
Stored the functions in R/ which is the standard place where R code
usually lives in a package which again, might be relevant if this
repository is to become a package at some point. Regardless this should
streamline the process of making changes to the many plots of the same
type in the report.

- Corrected some minor english typos in the report.

- Other minor changes/improvements.