Adding much faster and better compression options for RDS objects.
A major bottleneck in terms of time spent when working with the
pipeline, specifically with large files, is I/O. Saving one of the
pipeline objects, for a given step, can take upwards of 1 hour (default
"saveRDS"). The user can now choose an array of different options in
which to compress their RDS files. We still support the current way
("gzip") but now we can choose much better methods such as "lz4" and "zstd"
which are just much much faster with a small penalty if any regarding
size (see image attached on PR).

The functions "save_object" and "load_object" should, from now on, be the
interfaces through which we save and load RDS objects respectively.