Changed the preflight check
Previously, the pipeline would require you to start with scrna_phase_preprocess or singleton or by loading the data from a previous run. 

Now, you should be able to start with scrna_rawdata if you want to not use the phases.

This way, you can start a new run from scratch while only using parts of the preprocessing phase.